Challenge #8: Aggregate Consumer Purchases
Challenge Description
-------------------------
Aggregate Consumer Purchases:

For this week’s exercise we will look at customer purchase behavior to decide if we should offer a “Meal Deal” that would add a side and drink to a purchase of pizza or a burger. The incoming data is larger than usual for these exercises so I have packaged the workflow as an Alteryx Package. The link to the solution for last challenge #7 is HERE.

This week’s Objective:

In order to decide if we should start including a new "Meal Deal" on our menu we want to study the potential impact on recent transactions. Please identify the number and percentage of orders since July 1, 2013 which include the following categories of food: Pizza OR Burger along with a Side and Drink.

Summary of Data:
Point of Sale data includes the ticket level information, and the lookup table categorizes items into higher level food categories.

Hint:
Don't forget to join to the lookup table and filter by date.

As always we look forward to your feedback and suggestions!

UPDATE 01/18/2016:

The solution has been uploaded.

UPDATE 12/28/2016:

The challenge, text and solution have been updated.