Challenge #258: Relatively Old
Challenge Description:
------------------------
This week's challenge focuses on a group of relatives and their birthdays. Of the six family members, we have one person's date of birth and a series of clues to calculate the others. Provide the DOB for all family members and the day of the week s/he was born.

https://community.alteryx.com/t5/Weekly-Challenges/Challenge-258-Relatively-Old/td-p/729470