Challenge #324: Totals and Subtotals
Description:
-------------
When working with a traditional spreadsheet, it's relatively straightforward to summarize over rows and columns. But we often would like to do something similar in Designer! We also have the added feature to be able to create reports in a variety of formats.image324.png

For this challenge, find the total cost for all food items for each country as well as the total cost for each food item. Then create a report of your choice to show your findings.


Hint:
How to Generate and Format Totals and Subtotals 

https://community.alteryx.com/t5/Engine-Works/How-to-Generate-and-Format-Totals-and-Subtotals/ba-p/9... 

Challenge Link:
https://community.alteryx.com/t5/Weekly-Challenges/Challenge-324-Totals-and-Subtotals/td-p/956157